requirements.txt
anyio==3.6.2
-e git+https://github.com/pharmapsychotic/BLIP.git@5b58f59454039295a87c6628b1392e6b8752e911#egg=blip
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
-e git+https://github.com/openai/CLIP.git@d50d76daa670286dd6cacf3bcd80b5e4823fc8e1#egg=clip
clip-interrogator==0.1.4
colorama==0.4.6
contourpy==1.0.6
cycler==0.11.0
dnspython==2.2.1
email-validator==1.3.0
fairscale==0.4.4
fastapi==0.86.0
filelock==3.8.0
fonttools==4.38.0
ftfy==6.1.1
h11==0.14.0
httptools==0.5.0
huggingface-hub==0.10.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.6.2
numpy==1.23.4
orjson==3.8.1
packaging==21.3
Pillow==9.3.0
pycocoevalcap==1.2
pycocotools==2.0.6
pydantic==1.10.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
python-multipart==0.0.5
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
sacremoses==0.0.53
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
timm==0.4.12
tokenizers==0.10.3
torch==1.13.0
torchvision==0.14.0
tqdm==4.64.1
transformers==4.15.0
typing_extensions==4.4.0
ujson==5.5.0
urllib3==1.26.12
uvicorn==0.18.3
watchfiles==0.18.1
wcwidth==0.2.5
websockets==10.4
